Add skeleton for decoding tools and start working on database downloading function.
- I’m not sure if it’s currently possible to link Neurosynth/brainspell
annotations and coordinates to NeuroVault annotations and images. This
may require requesting additional features in NeuroVault and/or
brainspell.
- I don’t know if this is the best way to organize the decoding tools.
GCLDA, for example, is split across the different submodules based on
the type of decoding/encoding performed. Plus, I haven’t even begun to
figure out how we would integrate the dataset and model modules.
- Add some documentation to ALE fit method.