[REF] Update MetaResult and move general IBMA functions into new module (#152)
* Replace MetaResult with @tyarkoni's version

The only difference is that it takes in “mask” instead of “dataset”,
since retaining “dataset” at the point where MetaResults are generated
within the estimators would take a bit of refactoring at this point. I
can do it later though.

* Remove mask argument from most IBMA functions.

We still need to use it for the FSL-based ones, but hopefully we’ll be
able to write pure-Python versions of those at some point.

* Fix style problems.

* Split IBMA into IBMA and ESMA.

Also, move IBMA from its own folder to just a file.